+++ apps/user/src/components/form/ProgressBar/ProgressBar.tsx
@@ -1,6 +1,6 @@
 import { color, font } from '@maru/theme';
 import { flex } from '@maru/utils';
-import { useFormStepProvider } from '@/provider/useFormStepProvider';
+import { useFormStepProvider } from '@/hooks/provider/useFormStepProvider';
 import styled from 'styled-components';
 
 const PROGRESS_BAR_DATA = [
@@ -15,7 +15,6 @@ const PROGRESS_BAR_DATA = [
 /**
  * @TODO 다음페이지로 성공적으로 넘어갔을때 complete 처리를 해줘야합니다
  */
-
 
 const ProgressBar = () => {
     const { formStep, setFormStep } = useFormStepProvider();
