+++ apps/user/src/services/form/출신학교및학력/api.ts
@@ -1,7 +1,8 @@
 import { maru } from '@/apis/instance/instance';
+import { authorization } from '@/apis/token';
 
 export const getSchoolList = async (name: string) => {
-    const { data } = await maru.get(`/school?q=${name}`);
+    const { data } = await maru.get(`/school?q=${name}`, authorization());
 
     return data;
 };
